requirements.txt
acme==1.22.0
certbot==1.22.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==36.0.1
distro==1.6.0
idna==3.3
josepy==1.12.0
mypy==0.931
mypy-extensions==0.4.3
parsedatetime==2.6
pycparser==2.21
pyOpenSSL==21.0.0
pyRFC3339==1.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
tomli==2.0.0
types-cryptography==3.3.14
types-enum34==1.1.8
types-ipaddress==1.0.7
types-pyOpenSSL==21.0.3
types-PyYAML==6.0.3
types-requests==2.27.7
types-urllib3==1.26.7
typing-extensions==4.0.1
urllib3==1.26.8
zope.component==5.0.1
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
